add reward test
